server/requirements.txt
flask
pandas
pymssql
flask_cors
python-dotenv